# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flipzy API",
        "description": "Flipzy Property Management Platform API - Complete API documentation for property listings, messaging, analytics, and admin features.\n\n## Authentication\n\nThis API uses Laravel Passport for authentication. To authenticate:\n\n1. Use the `/register` or `/login` endpoint to get an access token\n2. Copy the `access_token` from the response\n3. Click the 'Authorize' button at the top of this page\n4. Enter: `Bearer YOUR_ACCESS_TOKEN` (replace YOUR_ACCESS_TOKEN with your actual token)\n5. Click 'Authorize' and then 'Close'\n\nAll protected endpoints will now use this token automatically.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "List all users (Admin only)",
                "description": "Get a paginated list of all users with optional filtering by role, search, and sorting",
                "operationId": "9680ccebed17e49ff174b2a620cc0cd5",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter by role",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "investor",
                                "wholesaler",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Get user details (Admin only)",
                "description": "Get detailed information about a specific user including roles, properties, and subscription",
                "operationId": "96a5f01127fa460057178398a6ddb53b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Update user (Admin only)",
                "description": "Update user information including name, email, role, and password",
                "operationId": "2255216aa7a93450befbfb0dd7dde469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "investor",
                                            "wholesaler",
                                            "admin"
                                        ]
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Delete user (Admin only)",
                "description": "Permanently delete a user account. Cannot delete your own account.",
                "operationId": "a42d951fbf6b2f8e25663e69c1b817d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "422": {
                        "description": "Cannot delete own account"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/suspend": {
            "post": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Suspend user account (Admin only)",
                "description": "Suspend a user account by removing email verification. Cannot suspend your own account.",
                "operationId": "b528be14e4e4df70fa51c75969e5d543",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User suspended successfully"
                    },
                    "422": {
                        "description": "Cannot suspend own account"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/activate": {
            "post": {
                "tags": [
                    "Admin - User Management"
                ],
                "summary": "Activate user account (Admin only)",
                "description": "Activate a suspended user account by verifying their email",
                "operationId": "9b7f56e8c44caa096965dc8dd3304a15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User activated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your access token in the format: Bearer YOUR_ACCESS_TOKEN\n\nGet your token by:\n- Registering a new account via POST /register\n- Logging in via POST /login\n\nThe token will be returned in the response as 'access_token'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}